Refactored new functionality Made drawing use inline font as well
